+++ src/components/form/text-field/styles.ts
@@ -1,4 +1,5 @@
 import { StyleDefType, TextStyle, ViewStyle } from '@alejandro.dev/rn-themizer'
+import { FontAwesomeIconStyle } from '@fortawesome/react-native-fontawesome'
 
 export type StyleGuide = {
     input: ViewStyle
@@ -6,6 +7,7 @@ export type StyleGuide = {
     label: TextStyle
     labelWrapper: ViewStyle
     root: ViewStyle
+    icon: FontAwesomeIconStyle
 }
 
 export type ParamsGiven = { focussed?: boolean; primary?: boolean; secondary?: boolean }
@@ -17,21 +19,37 @@ const styling: StyleDefType<StyleGuide, ParamsGiven> = ({
     params,
     palette
 }) => ({
+    icon: {
+        marginEnd: 10,
+        color: palette.primaryColor,
+        ...applyIf(params?.secondary, {
+            color: palette.secondaryColor
+        })
+    },
     input: {
         height: '100%',
         fontSize: fromVars('textField.input.fontSize', 18),
-        paddingHorizontal: fromVars('textField.input.paddingHorizontal', 10)
+        paddingHorizontal: fromVars('textField.input.paddingHorizontal', 10),
+        flex: 1
     },
     inputWrapper: {
+        alignItems: 'center',
         borderWidth: fromVars('textField.inputWrapper.borderWidth', 1),
         borderRadius: fromVars('textField.inputWrapper.borderRadius', 5),
         borderColor: fromPalette('textField.borderColor', '#90a4ae'),
         height: fromVars('textField.inputWrapper.height', 40),
         backgroundColor: fromPalette('textField.inputWrapper.bgColor', '#FFF'),
+        flexDirection: 'row',
         ...applyIf(params?.focussed, {
             borderWidth: fromVars('textField.inputWrapper.borderWidthFocussed', 2),
-            borderColor: fromPalette('primaryColor', '#90a4ae'),
-            shadowColor: fromPalette('primaryColor', '#90a4ae'),
+            borderColor: fromPalette(
+                params?.secondary ? 'secondaryColor' : 'primaryColor',
+                '#90a4ae'
+            ),
+            shadowColor: fromPalette(
+                params?.secondary ? 'secondaryColor' : 'primaryColor',
+                '#90a4ae'
+            ),
             shadowOffset: {
                 width: 0,
                 height: 2