+++ src/components/form/text-field/types.ts
@@ -1,3 +1,4 @@
+import { IconNameType } from '../../icon/types'
 export interface TextFieldProps {
     label?: string
     secondary?: boolean
@@ -7,4 +8,5 @@ export interface TextFieldProps {
     onFocus?: () => void
     placeholder?: string
     value?: string
+    icon?: IconNameType
 }